+++ Bybit.Api.Examples/Usings.cs
@@ -1,5 +1,6 @@
 ﻿global using ApiSharp.Authentication;
 global using Bybit.Api.Account;
+global using Bybit.Api.Affiliate;
 global using Bybit.Api.Asset;
 global using Bybit.Api.Enums;
 global using Bybit.Api.Margin;